refactor: extract `requirementProvidedValues` config to its own file
