+++ src/components/create-guild/Requirements/components/ProvidedValueDisplay.tsx
@@ -14,7 +14,7 @@ import {
 } from "@chakra-ui/react"
 import { Info, Lightning, Question } from "@phosphor-icons/react"
 import Card from "components/common/Card"
-import { REQUIREMENT_PROVIDED_VALUES } from "requirements/requirements"
+import { REQUIREMENT_PROVIDED_VALUES } from "requirements/requirementProvidedValues"
 import { Requirement } from "types"
 
 const ProvidedValueDisplay = ({
