+++ src/requirements/Airdrop/providedValue/AirdropProvidedValue.tsx
@@ -2,7 +2,7 @@ import { HStack, Text, useDisclosure } from "@chakra-ui/react"
 import { ArrowSquareIn } from "@phosphor-icons/react"
 import SnapshotModal from "components/[guild]/leaderboard/Snapshots/SnapshotModal"
 import Button from "components/common/Button"
-import { ProvidedValueDisplayProps } from "requirements"
+import type { ProvidedValueDisplayProps } from "requirements/requirementProvidedValues"
 
 const AirdropProvidedValue = ({ requirement }: ProvidedValueDisplayProps) => {
   const { isOpen, onOpen, onClose } = useDisclosure()
