+++ src/requirements/Farcaster/FarcasterForm.tsx
@@ -2,7 +2,7 @@ import { FormControl, FormLabel, Stack } from "@chakra-ui/react"
 import ControlledSelect from "components/common/ControlledSelect"
 import { useFormContext, useWatch } from "react-hook-form"
 import { RequirementFormProps, RequirementType } from "requirements"
-import { PROVIDER_TYPES } from "requirements/requirements"
+import { PROVIDER_TYPES } from "requirements/requirementProvidedValues"
 import FarcasterCastHash from "./components/FarcasterCastHash"
 import FarcasterChannel from "./components/FarcasterChannel"
 import FarcasterTextToInclude from "./components/FarcasterTextToInclude"
