+++ src/requirements/GitcoinPassport/GitcoinPassportForm.tsx
@@ -8,7 +8,7 @@ import {
 import ControlledSelect from "components/common/ControlledSelect"
 import { useFormContext, useWatch } from "react-hook-form"
 import { RequirementFormProps, RequirementType } from "requirements"
-import { PROVIDER_TYPES } from "requirements/requirements"
+import { PROVIDER_TYPES } from "requirements/requirementProvidedValues"
 import parseFromObject from "utils/parseFromObject"
 import Score from "./components/Score"
