+++ src/requirements/Guild/GuildForm.tsx
@@ -4,7 +4,7 @@ import FormErrorMessage from "components/common/FormErrorMessage"
 import { useFormContext, useWatch } from "react-hook-form"
 import { RequirementFormProps, RequirementType } from "requirements"
 import GuildSelect from "requirements/common/GuildSelect"
-import { PROVIDER_TYPES } from "requirements/requirements"
+import { PROVIDER_TYPES } from "requirements/requirementProvidedValues"
 import parseFromObject from "utils/parseFromObject"
 import GuildAdmin from "./components/GuildAdmin"
 import MinGuilds from "./components/MinGuilds"
