+++ src/requirements/Lens/LensForm.tsx
@@ -2,7 +2,7 @@ import { Divider, FormControl, FormLabel, Stack } from "@chakra-ui/react"
 import ControlledSelect from "components/common/ControlledSelect"
 import { useFormContext, useWatch } from "react-hook-form"
 import { RequirementFormProps, RequirementType } from "requirements"
-import { PROVIDER_TYPES } from "requirements/requirements"
+import { PROVIDER_TYPES } from "requirements/requirementProvidedValues"
 import LensAction from "./components/LensAction"
 import LensPostInput from "./components/LensPostInput"
 import LensProfileSelect from "./components/LensProfileSelect"
