+++ src/requirements/Points/providedValue/PointsRankProvidedValue.tsx
@@ -1,5 +1,5 @@
 import useGuildPlatform from "components/[guild]/hooks/useGuildPlatform"
-import { ProvidedValueDisplayProps } from "requirements"
+import type { ProvidedValueDisplayProps } from "requirements/requirementProvidedValues"
 
 const PointsRankProvidedValue = ({ requirement }: ProvidedValueDisplayProps) => {
   const { guildPlatform } = useGuildPlatform(requirement?.data?.guildPlatformId)
