+++ src/requirements/Snapshot/SnapshotForm.tsx
@@ -3,7 +3,7 @@ import ControlledSelect from "components/common/ControlledSelect"
 import FormErrorMessage from "components/common/FormErrorMessage"
 import { useFormContext, useFormState, useWatch } from "react-hook-form"
 import { RequirementFormProps, RequirementType } from "requirements"
-import { PROVIDER_TYPES } from "requirements/requirements"
+import { PROVIDER_TYPES } from "requirements/requirementProvidedValues"
 import parseFromObject from "utils/parseFromObject"
 import FollowSince from "./components/FollowSince"
 import MajorityVotes from "./components/MajorityVotes"
