+++ src/requirements/Uniswap/providedValue/PositionsProvidedValue.tsx
@@ -1,6 +1,6 @@
 import { HStack, Skeleton, Text } from "@chakra-ui/react"
 import DataBlock from "components/common/DataBlock"
-import { ProvidedValueDisplayProps } from "requirements"
+import type { ProvidedValueDisplayProps } from "requirements/requirementProvidedValues"
 import { Chains } from "wagmiConfig/chains"
 import { useSymbolsOfPair } from "../hooks/useSymbolsOfPair"
