+++ src/requirements/WalletActivity/WalletActivityForm.tsx
@@ -4,7 +4,7 @@ import FormErrorMessage from "components/common/FormErrorMessage"
 import { useFormContext, useWatch } from "react-hook-form"
 import { RequirementFormProps, RequirementType } from "requirements"
 import ChainPicker from "requirements/common/ChainPicker"
-import { PROVIDER_TYPES } from "requirements/requirements"
+import { PROVIDER_TYPES } from "requirements/requirementProvidedValues"
 import { SelectOption } from "types"
 import parseFromObject from "utils/parseFromObject"
 import { Chain } from "wagmiConfig/chains"
