+++ src/requirements/WalletActivity/providedValue/TxCountProvidedValue.tsx
@@ -1,6 +1,6 @@
 import { HStack, Text, useColorModeValue } from "@chakra-ui/react"
 import RequirementChainIndicator from "components/[guild]/Requirements/components/RequirementChainIndicator"
-import { ProvidedValueDisplayProps } from "requirements"
+import type { ProvidedValueDisplayProps } from "requirements/requirementProvidedValues"
 
 const TxCountProvidedValue = ({ requirement }: ProvidedValueDisplayProps) => {
   const tagBg = useColorModeValue("blackAlpha.100", "blackAlpha.300")
